Changed the way Labrynths are saved.
Moved all the code which handles saving LabrynthDetails from EntityPMWitchLabrynthEntrance to LabrynthManager. Basically, we're now saving LabrynthDetails into a new custom file in the player's world, rather than piggy-backing off of each EntityPMWitchLabrynthEntrance's NBT data. Now each EntityPMWitchLabrynthEntrance only stores the ID number of the dimension it points to. The reason I changed this was because I Dimensions are loaded before Entitys, so we need to make sure that the dimension actually exists before the entity loads. Added some code to PMEventHandler which saves and loads the data. I've also added a bunch of methods to handle this stuff.