Moved some methods around.
Moved the onEntityItemPickup and onEntityLivingAttacked methods from PMDataTracker to PMEventHandler. This fixes a problem with them where they wouldn't ever run (due to the fact that you have to register event handlers, not just subscribe to events). Also modified the methods slightly to make them work outside of the context of PMDataTracker. Added some getter methods to PMDataTracker.