Fixed a mistake I made.
Added a line which increments a counter that is supposed to keep track of the current tracker # that we are loading (I forgot that line last time because I'm an idiot). Added a method which returns whether or not any tracker is considered dirty (needs to be saved to a file). Removed the saveAllTrackers method (writePersistentFile handles everything anyways).