Rewrote the bji command.
Moved all the code which handles this into a separate class (so that I may use it in a separate project some day). Added a special class designed to help BasicJavaInterpreter interface with Minecraft. Attempted to fix the issue of being unable to chain together methods. Added support for raw-values instead of just calling methods (only works with ints and Strings).