Some minor changes and one major change.
Changed loadTagData to return a boolean as status. Added an ID variable (really, it's only going to be useful for some minor printing issues. Added a NBTTagCompound field, which is where all data is going to be saved, rather than piggy-backing the data off of the host Entity's data. I should have had this done a while ago, especially the writing of data already does this, I just forgot to do it for the reading of data too.