Added some debuging code and a new AI routine.
Added a special method specifically for randomizing all data about a witch. At the moment, this method is called on creation, but this should be fixed as the method should only happen when a witch is spawned without a home labrynths (i.e from a spawn egg). This method is purely for testing purposes since the rest of the puella magi system isn't finished yet. Added a new AI task to hopefully replace the old RandomTeleportPlayerOrVillager routine.