Fixed a really stupid mistake.
Fixed an error that would cause a NullPointerException on startup, which was caused when createChunkGenerator returned null, leading to it getting passed around all over the place before finally resulting in a NullPointerException in some very deeply buried bit of code (ChunkProviderServer#unloadQueuedChunks@Line409 to be specific). After tracking down the problem, I realized that createChunkGenerator gets assigned to currentChunkProvider in ChunkProviderServer, and this field can NEVER be null, otherwise you get this same issue. Added a comment.