Added more methods.
Added a method to request a WorldServer object belonging to a specific dimension. If this dimension hasn't been registered, then null is returned and if the dimension hasn't been initialized then it is and the resulting world is returned. Added a method to transport the player from one location to another, in terms of x, x,y, x,y,z, and x,y,z,dimension. Added a comment.