Attempted to fix some issues when randomizing labrynths.
Moved call to createRandomizedWitch outside of the constructor in an attempt to fix a bug where it would be called twice (hasn't fixed it yet). Changed the NBT Tag for the linked dimension ID from 'Linked Dimension ID' to 'LINKED_ID'. Added some code which allows EntityPMWitchLabrynthEntrance objects to be damaged if a config value is true. Added some code to onDeath in an attempt to clear data when we don't need it anymore. Added some code to createRandomizedWitch in an attempt to curb the runaway creation of new dimensions (doesn't quit work yet either).