Attempted to fix various problems related to the creation and deletion of labrynths.
Changed the NBTCompound structure that we save data as to more closely match PlayerDataTrackerManager and to fix issues where nothing would be saved. Added a field to map Labrynth entrances (by UUID) to their linked labrynth's details. Added getters for this new field. Deprecated a method. Moved some method calls around (too many to list and describe here). Added code to deleteData(LabrynthDetails) to make sure that a world is not loaded before unregistering it from DimensionManager. Rewrote storeLabrynthEntranceToRetrieveLater(...) so that we don't have to worry about the array becoming too full if left running for too long. Removed the call to checkAndUpdate from retrieveEntrance.