Fixed some runtime and compile-time issues. Made a minor improvement.
Fixed some issues that occurred with the last two commits (a few NullPointerExceptions, and a compile error). Added some code to PlayerDataTrackerManager so that it attempts to clean up the set of trackers. Fixed an issue where PMDataTracker wasn't saving all the data it was supposed to, causing issues when loading it.