Completed a TODO and improved some old code.
Fixed the code which fires an event to create a new witch if the player is done transforming. Fixed the code which creates a witch when a MadokaMagicaCreateWitchEvent is fired by updating the calls from old, deprecated methods to new ones, as well as adding some code which sets the position and dimension of both the newly created witch and labrynth entrance. Fixed a possible logic error regarding some math.